requirements.txt
streamlit
langchain
langchain-google-genai
langchain-mcp-adapters
langgraph
langchain-openai
openai
python-dotenv
